# Conflicts:
#	StockSip.Platform.API/PaymentAndSubscription/Application/Internal/QueryService/AccountQueryService.cs
#	StockSip.Platform.API/PaymentAndSubscription/Domain/Model/Aggregates/Account.cs
#	StockSip.Platform.API/PaymentAndSubscription/Domain/Repositories/IAccountRepository.cs
#	StockSip.Platform.API/PaymentAndSubscription/Domain/Services/IAccountQueryService.cs
#	StockSip.Platform.API/PaymentAndSubscription/Infrastructure/Configuration/Extensions/ModelBuilderExtensions.cs
#	StockSip.Platform.API/PaymentAndSubscription/Infrastructure/Persistence/Repositories/AccountRepository.cs
#	StockSip.Platform.API/PaymentAndSubscription/Interfaces/REST/AccountsController.cs
#	StockSip.Platform.API/Program.cs
#	StockSip.Platform.API/Shared/Infrastructure/Persistence/EFC/Configuration/AppDbContext.cs

+++ StockSip.Platform.API/InventoryManagement/Interfaces/REST/Resources/UpdateProductResource.cs
@@ -6,6 +6,6 @@
 public record UpdateProductResource(string Name, 
                                     string BrandName, 
                                     string LiquorType, 
-                                    double UnitPriceAmount, 
+                                    decimal UnitPriceAmount, 
                                     int MinimumStock, 
                                     IFormFile? Image);
